requirements.txt
# Core dependencies for the chatbot
PyJWT
python-jose
passlib[bcrypt]
python-dotenv==1.0.1
langchain-core==0.3.66
langchain-community==0.3.26
langchain==0.3.26
langchain-openai==0.2.8
openai==1.93.0
tiktoken==0.9.0

# FastAPI and web dependencies
fastapi==0.110.2
uvicorn==0.29.0
requests==2.31.0

# Database dependencies
sqlalchemy==2.0.30
psycopg2-binary==2.9.9

# Vector store and ML dependencies
pinecone-client==3.2.2
sentence-transformers==2.6.1
numpy==1.26.4
torch==2.3.1

# Data processing
pandas==2.2.2

# Web scraping dependencies (optional for deployment)
selenium==4.20.0
webdriver-manager==4.0.1

# Additional utilities
python-multipart==0.0.9
pydantic==2.11.7

# Production dependencies
gunicorn==21.2.0

# Testing tools
httpx==0.27.0
pytest==8.1.1